More refactoring, condensed each parse file pair (i.e. Vehicle and VehicleParsing) into one file
