+++ Parse/Route.swift
@@ -8,27 +8,8 @@
 
 import Foundation
 
-/*
- Route struct:
- color: String
- created: String (time object)
- description: String
- enabled: Bool (int rep)
- id: Int
- name: String
- points: Array:
-    latitude: Double
-    longitude: Double
- stop_ids: Ints? (empty)
- updated: String (time object)
- width: Int
- 
- ^^ order in which webpage displays it
- */
-
 struct Route {
     
-    //default init
     var color = ""
     var created = ""
     var description = ""
@@ -40,6 +21,58 @@ struct Route {
     var updated = ""
     var width = 0
     
+    init?(json: NSDictionary) {
+        //need to add the points to the route
+        var pointsList:[Point] = [];
+        
+        //for each different value of each route
+        for (key, value) in json {
+            switch key as? NSString {
+                
+            //catching different variables for a route
+            case "color":
+                self.color = (value as! String)
+            case "created":
+                self.created = (value as! String)
+            case "description":
+                self.description = (value as! String)
+            case "enabled":
+                self.enabled = (value as! Bool)
+            case "id":
+                self.id = (value as! Int)
+            case "name":
+                self.name = (value as! String)
+            case "stop_ids":
+                self.stop_ids = (value as! [Int])
+            case "updated":
+                self.updated = (value as! String)
+            case "width":
+                self.width = (value as! Int)
+            case "points":
+                //need to parse the coordinates separately since
+                //nested dictionary inside a dictionary
+                for loc in (value as! NSArray) {
+                    var lat: Double = 0
+                    var long: Double = 0
+                    for piece in (loc as! NSDictionary) {
+                        if ((piece.key as! String) == "latitude") {
+                            lat = piece.value as! Double
+                        } else {
+                            long = piece.value as! Double
+                        }
+                    }
+                    //append the points to the array
+                    pointsList.append(Point(latitude: lat, longitude: long))
+                }
+            default:
+                //should never go off
+                print("\(key)")
+            }
+        }
+        //points set at the end
+        self.points = pointsList
+        print("Finished JSON initialization for route \(self.id)")
+    }
     
     func printRoute(){
         print("Color: \(self.color)")
@@ -49,7 +82,7 @@ struct Route {
         print("Id: \(self.id)")
         print("Name: \(self.name)")
         printPoints()
-        print("Stop_IDs: \(self.stop_ids)")
+        print("Stop IDs: \(self.stop_ids)")
         print("Updated: \(self.updated)")
         print("Width: \(self.width)")
     }
