+++ Parse/Vehicle.swift
@@ -17,6 +17,30 @@ struct Vehicle {
     var enabled = false
     var tracker_id = ""
     
+    init?(json: NSDictionary) {
+        // Catch vehicle data
+        for (key, value) in json {
+            switch key as? NSString {
+            case "id":
+                self.id = (value as! Int)
+            case "name":
+                self.name = (value as! String)
+            case "created":
+                self.created = (value as! String)
+            case "updated":
+                self.updated = (value as! String)
+            case "enabled":
+                self.enabled = (value as! Bool)
+            case "tracker_id":
+                self.tracker_id = (value as! String)
+            default:
+                // This should never happen
+                print("\(key)")
+            }
+        }
+        print("Finished JSON initialization for vehicle \(self.id)")
+    }
+    
     func printVehicle() {
         print("ID: \(self.id)")
         print("Name: \(self.name)")
